Changes as per PR review.
Put vue tests in a separate file and make the html template an instance.
