+++ jigna/template.py
@@ -5,28 +5,13 @@
 # All right reserved.
 #
 
+from textwrap import dedent
+
 # Enthought library.
 from traits.api import Bool, HasTraits, Str, Property, Tuple, Int
 
+
 #### HTML template ############################################################
-
-DOCUMENT_HTML_TEMPLATE = """
-<html ng-app='jigna'>
-  <head>
-    <script type="text/javascript" src="/jigna/jigna.js"></script>
-    <script type="text/javascript">
-        jigna.initialize({{async: {async}}});
-    </script>
-
-    {head_html}
-
-  </head>
-
-  <body>
-    {body_html}
-  </body>
-</html>
-"""
 
 class Template(HasTraits):
     """ Encapsulation of the HTML/AngularJS template which can be rendered by
@@ -57,9 +42,15 @@ class Template(HasTraits):
     #: The inner HTML for the *head* of the view's document.
     head_html = Str
 
+    #: The foot HTML for the *head* of the view's document.
+    foot_html = Str
+
     #: The file which contains the html. `html_file` takes precedence over
     #: `body_html` and `head_html`.
     html_file = Str
+
+    #: The HTML template used for this.
+    html_template = Str
 
     #: The HTML for the entire document.
     #:
@@ -89,9 +80,10 @@ class Template(HasTraits):
         # ...otherwise, create the template out of body and head htmls
         else:
             async = 'true' if self.async else 'false'
-            html = DOCUMENT_HTML_TEMPLATE.format(
+            html = self.html_template.format(
                 body_html = self.body_html,
                 head_html = self.head_html,
+                foot_html = self.foot_html,
                 async     = async,
             )
 
@@ -99,3 +91,24 @@ class Template(HasTraits):
 
     def _set_html(self, html):
         self._html = html
+
+    def _html_template_default(self):
+        return dedent("""
+        <html ng-app='jigna'>
+          <head>
+            <script type="text/javascript" src="/jigna/jigna.js"></script>
+            <script type="text/javascript">
+                jigna.initialize({{async: {async}}});
+            </script>
+
+            {head_html}
+
+          </head>
+
+          <body>
+            {body_html}
+          </body>
+
+            {foot_html}
+        </html>
+        """)
